fix: Move common methods to transfromer.py
Move common methods to transfromer.py
to have less copy pasted code.
Added more debug messages to transformation procedure.
